moved batch info and block info to state diffs
