+++ zkos/recovery/src/main.rs
@@ -1,24 +1,22 @@
 use std::{collections::HashMap, str::FromStr};
 
 use alloy::{
-    consensus::Transaction,
-    primitives::{Address, B256, U256, address},
+    primitives::{Address, B256},
     providers::{Provider, ProviderBuilder},
     rpc::types::Filter,
     sol,
-    sol_types::{SolCall, SolEvent}, // for ABI-safe decoding of the commit function
+    sol_types::SolEvent, // for ABI-safe decoding of the commit function
 };
 use anyhow::{Context, Result, bail};
-use blake2::{Blake2s256, Digest};
 use clap::Parser;
 use tracing::level_filters::LevelFilter;
 use tracing_subscriber::EnvFilter;
 
 use crate::{
     chain_genesis::get_genesis_upgrade,
-    state::{BlockchainState, LocalTree},
+    state::BlockchainState,
     state_genesis::init_genesis,
-    statediffs::{StateDiff, ValueDiff},
+    statediffs::{BatchInfo, BlockInfo},
 };
 
 pub mod bytecodes;
@@ -271,23 +269,6 @@ async fn main() -> Result<()> {
     Ok(())
 }
 
-pub struct BatchInfo {
-    pub batch_number: u64,
-    pub batch_hash: B256,
-    pub commitment: B256,
-    pub calldata: Vec<u8>,
-    pub stored: StoredBatchInfo,
-    pub commits: Vec<CommitBatchInfoZKsyncOS>,
-    // Block hash, state diffs, logs
-    pub blocks_data: Vec<BlockInfo>,
-}
-
-pub struct BlockInfo {
-    pub block_hash: B256,
-    pub state_diffs: Vec<StateDiff>,
-    pub logs: Vec<statediffs::Log>,
-}
-
 async fn get_commit_batches_from_range<P: Provider + Clone>(
     provider: &P,
     address: Address,
@@ -326,245 +307,10 @@ async fn get_commit_batches_from_range<P: Provider + Clone>(
             data: lg.data().clone(),
         };
         if let Ok(ev) = BlockCommit::decode_log(&prim_log) {
-            let batch = ev.batchNumber;
-            let batch_hash = ev.batchHash;
-            let commitment = ev.commitment;
-
-            let calldata = tx.input().clone();
-
-            tracing::debug!(
-                "{{\"block\":{},\"tx\":\"0x{}\",\"batchNumber\":{},\"batchHash\":\"0x{}\",\"commitment\":\"0x{}\"}}",
-                lg.block_number.unwrap_or_default(),
-                hex::encode(tx_hash.as_slice()),
-                batch,
-                hex::encode(batch_hash.as_slice()),
-                hex::encode(commitment.as_slice()),
-                //hex::encode(&calldata),
-            );
-
-            let commit_call = commitBatchesSharedBridgeCall::abi_decode(&calldata).unwrap();
-
-            let commit_data_without_prefix = &commit_call._commitData[1..]; // skip the first byte (version)
-
-            let decode_params =
-                __decodeParamsCall::abi_decode_raw(commit_data_without_prefix).unwrap();
-
-            let (stored, commits) = (decode_params._0, decode_params._1);
-
-            let tmp = &commits[0];
-            assert_eq!(1, commits.len());
-
-            let blocks_data = parse_da_input(&tmp.operatorDAInput).unwrap();
-
-            let batch_number: u64 = batch.try_into().unwrap();
-            results.insert(
-                batch_number,
-                BatchInfo {
-                    batch_number,
-                    batch_hash,
-                    commitment,
-                    calldata: calldata.to_vec(),
-                    stored,
-                    commits,
-                    blocks_data,
-                },
-            );
+            let batch_info = BatchInfo::parse(ev, tx);
+            results.insert(batch_info.batch_number, batch_info);
         }
     }
 
     Ok(results)
-}
-
-pub fn parse_da_input(input: &[u8]) -> Result<Vec<BlockInfo>> {
-    // first 32 bytes should be 0, the next 32 is some keccak.
-    if input.len() < 64 {
-        tracing::error!("DA input too short: {}", input.len());
-        return Err(anyhow::anyhow!("DA input too short"));
-    }
-    // not sure what this prefix is..
-    let prefix = &input[0..32];
-    let pubdata_hash = &input[32..64];
-    if prefix.iter().any(|&b| b != 0) {
-        tracing::error!("DA input prefix not zero: {:x?}", prefix);
-        return Err(anyhow::anyhow!("DA input prefix not zero"));
-    }
-    tracing::debug!("pubdata input hash: 0x{}", hex::encode(pubdata_hash));
-    let blob_count = &input[64];
-    // for calldata, blobcount should be 1.
-
-    tracing::debug!("blob count: {}", blob_count);
-    let mut offset = 65;
-    // another 32 bytes that should be 0.
-    if input.len() < offset + 32 {
-        tracing::error!("DA input too short for second zero: {}", input.len());
-        return Err(anyhow::anyhow!("DA input too short for second zero"));
-    }
-    let mid = &input[offset..offset + 32];
-    if mid.iter().any(|&b| b != 0) {
-        tracing::error!("DA input mid not zero: {:x?}", mid);
-        return Err(anyhow::anyhow!("DA input mid not zero"));
-    }
-    offset += 32;
-
-    let calldata_type = &input[offset];
-    tracing::debug!("calldata type: {}", calldata_type);
-    assert_eq!(&0, calldata_type); // we only handle calldata type 0
-
-    offset += 1;
-
-    // remaining bytes:
-
-    let mut results = vec![];
-    loop {
-        tracing::debug!("parsing block {}", results.len());
-        let (consumed, block_header_hash, state_diff, logs) = parse_block_da(&input[offset..])?;
-        tracing::debug!(
-            "offset: {} consumed: {} input len: {}",
-            offset,
-            consumed,
-            input.len()
-        );
-        offset += consumed;
-        results.push(BlockInfo {
-            block_hash: block_header_hash,
-            state_diffs: state_diff,
-            logs,
-        });
-        // TODO: seems that last 32 bytes are 0s..
-        if offset + 32 >= input.len() {
-            break;
-        }
-    }
-    let blob_commitment = B256::from_slice(&input[offset..offset + 32]);
-    tracing::debug!("blob commitment: {:?}", blob_commitment);
-    assert_eq!(blob_commitment, B256::ZERO);
-
-    Ok(results)
-}
-
-pub fn parse_block_da(input: &[u8]) -> Result<(usize, B256, Vec<StateDiff>, Vec<statediffs::Log>)> {
-    let pubdata = input;
-    tracing::debug!("pubdata len: {}", pubdata.len());
-
-    // now for pubdata itself.
-
-    // First 32 should be some hash.
-    if pubdata.len() < 32 {
-        tracing::error!("pubdata too short for hash: {}", pubdata.len());
-        return Err(anyhow::anyhow!("pubdata too short for hash"));
-    }
-    // This is the 'current_block_hash' from io_subsystem.rs 'finish'
-    let block_header_hash = B256::from_slice(&pubdata[0..32]);
-    tracing::debug!("block header hash: 0x{}", hex::encode(block_header_hash));
-
-    let (state_diff_offset, state_diff) = StateDiff::new_from_stream(&pubdata[32..]);
-
-    let remaining = &pubdata[32 + state_diff_offset as usize..];
-
-    // u32 for logs length
-    if remaining.len() < 4 {
-        tracing::error!("pubdata too short for logs len: {}", remaining.len());
-        return Err(anyhow::anyhow!("pubdata too short for logs len"));
-    }
-    let logs_len = u32::from_be_bytes(
-        remaining[0..4]
-            .try_into()
-            .expect("slice with incorrect length"),
-    );
-    tracing::debug!("pubdata logs len: {}", logs_len);
-
-    let mut offset = 4;
-
-    let mut logs = Vec::new();
-
-    for _ in 0..logs_len {
-        let (consumed, log) = statediffs::Log::new_from_stream(&remaining[offset..]);
-
-        logs.push(log);
-        offset += consumed as usize;
-    }
-
-    let messages_len: u32 = if remaining.len() < offset + 4 {
-        tracing::error!("pubdata too short for messages len: {}", remaining.len());
-        return Err(anyhow::anyhow!("pubdata too short for messages len"));
-    } else {
-        u32::from_be_bytes(
-            remaining[offset..offset + 4]
-                .try_into()
-                .expect("slice with incorrect length"),
-        )
-    };
-    offset += 4;
-
-    tracing::debug!("pubdata messages len: {}", messages_len);
-
-    if messages_len > 0 {
-        for _ in 0..messages_len {
-            let len = u32::from_be_bytes(
-                remaining[offset..offset + 4]
-                    .try_into()
-                    .expect("slice with incorrect length"),
-            );
-            offset += 4;
-            tracing::trace!("message len: {}", len);
-            offset += len as usize;
-        }
-    }
-
-    tracing::debug!("pubdata remaining len: {}", remaining.len() - offset);
-
-    Ok((
-        offset + 32 + state_diff_offset as usize,
-        block_header_hash,
-        state_diff,
-        logs,
-    ))
-}
-
-pub fn address_to_b256(address: &Address) -> B256 {
-    let mut extended_address = [0u8; 32];
-    extended_address[12..].copy_from_slice(&address.0.0);
-    B256::from(extended_address)
-}
-
-pub fn derive_properties_storage_address(address: &Address) -> B256 {
-    let account_properties_address = address!("0000000000000000000000000000000000008003");
-
-    let mut hasher = Blake2s256::new();
-    hasher.update(address_to_b256(&account_properties_address));
-    hasher.update(address_to_b256(address));
-
-    let hash = hasher.finalize();
-    B256::from_slice(&hash)
-}
-
-pub fn u256_to_b256(value: &U256) -> B256 {
-    let bytes = value.to_be_bytes();
-
-    B256::from(bytes)
-}
-
-pub fn b256_to_u256(value: &B256) -> U256 {
-    U256::from_be_bytes(value.0)
-}
-
-pub fn apply_value_diff(tree: &mut LocalTree, key: B256, diff: &ValueDiff) {
-    match diff {
-        ValueDiff::Nothing(v) => {
-            tree.add_entry(key, u256_to_b256(v));
-        }
-        ValueDiff::Add(v) => tree.add_entry(
-            key,
-            u256_to_b256(&b256_to_u256(&tree.get_value(key)).wrapping_add(*v)),
-        ),
-        ValueDiff::Sub(v) => {
-            tree.add_entry(
-                key,
-                u256_to_b256(&b256_to_u256(&tree.get_value(key)).wrapping_sub(*v)),
-            );
-        }
-        ValueDiff::Transform(v) => {
-            tree.add_entry(key, u256_to_b256(v));
-        }
-    };
 }
